+++ cli/memory_custodian/__init__.py
@@ -1,4 +1,4 @@
 """MemoryCustodian CLI package."""
 
-__version__ = "0.4.0"
+__version__ = "0.4.1"
 __protocol_version__ = "0.4"